:recycle: api: move ComplementaryCertificationScoringCriteria to its domain
