+++ api/tests/tooling/domain-builder/factory/certification/evaluation/build-complementary-certification-scoring-criteria.js
@@ -1,4 +1,4 @@
-import { ComplementaryCertificationScoringCriteria } from '../../../../src/shared/domain/models/ComplementaryCertificationScoringCriteria.js';
+import { ComplementaryCertificationScoringCriteria } from '../../../../../../src/certification/evaluation/domain/models/ComplementaryCertificationScoringCriteria.js';
 
 function buildComplementaryCertificationScoringCriteria({
   complementaryCertificationCourseId = 123,
